move macOS mods to separate file
