+++ plugins/index.js
@@ -3,56 +3,9 @@ const { withMod } = require("@expo/config-plugins");
 
 /**
  * @typedef {import("@expo/config-plugins").ExportedConfig} ExportedConfig
- * @typedef {import("@expo/config-plugins").ExportedConfigWithProps} ExportedConfigWithProps
  * @typedef {import("@expo/config-plugins").Mod} Mod
  * @typedef {import("@expo/config-plugins").ModConfig} ModConfig
- * @typedef {ExportedConfigWithProps & { macos?: { infoPlist?: Record<string, unknown> }}} ExportedConfigWithPropsMac
  */
-
-const macosPlatform = /** @type {keyof ModConfig} */ ("macos");
-
-/** @type {Record<string, (config: ExportedConfig, mod: Mod) => ExportedConfig>} */
-const macos = {
-  withReactNativeHost: (config, action) => {
-    return withMod(config, {
-      platform: macosPlatform,
-      mod: "reactNativeHost",
-      action,
-    });
-  },
-  // https://github.com/expo/expo/blob/sdk-51/packages/%40expo/config-plugins/src/plugins/ios-plugins.ts#L101
-  withAppDelegate: (config, action) => {
-    return withMod(config, {
-      platform: macosPlatform,
-      mod: "appDelegate",
-      action,
-    });
-  },
-  // https://github.com/expo/expo/blob/sdk-51/packages/%40expo/config-plugins/src/plugins/ios-plugins.ts#L116
-  withInfoPlist: (config, action) => {
-    return withMod(config, {
-      platform: macosPlatform,
-      mod: "infoPlist",
-      async action(cfg) {
-        /** @type {ExportedConfigWithPropsMac} */
-        const config = await action(cfg);
-        if (!config.macos) {
-          config.macos = {};
-        }
-        config.macos.infoPlist = config.modResults;
-        return config;
-      },
-    });
-  },
-  // https://github.com/expo/expo/blob/sdk-51/packages/%40expo/config-plugins/src/plugins/ios-plugins.ts#L173
-  withXcodeProject: (config, action) => {
-    return withMod(config, {
-      platform: macosPlatform,
-      mod: "xcodeproj",
-      action,
-    });
-  },
-};
 
 /**
  * Provides the `ReactNativeHost` file for modification.
@@ -82,6 +35,6 @@ function withSceneDelegate(config, action) {
   });
 }
 
-exports.macos = macos;
+exports.macos = require("./macos");
 exports.withReactNativeHost = withReactNativeHost;
 exports.withSceneDelegate = withSceneDelegate;
